requirements.txt
asgiref==3.8.1
certifi==2025.4.26
chardet==5.2.0
charset-normalizer==3.4.2
dateutils==0.6.12
Django==5.2.3
ecdsa==0.19.1
idna==3.10
pillow==11.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
qrcode==8.2
reportlab==4.4.1
requests==2.32.4
six==1.17.0
sqlparse==0.5.3
urllib3==2.4.0
